uast/yaml: rename the package to match an import path
