+++ uast/uastyaml/uastyaml_test.go
@@ -1,4 +1,4 @@
-package uastyml
+package uastyaml
 
 import (
 	"testing"
